+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **Added:** Additional configuration options:
     - `reminder_threshold`: Sets the number of days at/below which Colette will print a reminder about a task with an approaching due date.
 - **Fixed:** Tasks with passed due dates being printed under reminders instead of passed due date notifications.
+- **Fixed:** `autoremove` command was only removing single entries instead of multiple.
 
 ## v0.2.1 (2024-06-26)